backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-multipart>=0.0.9
numpy>=1.26.0
opencv-python-headless>=4.10.0.84
google-cloud-vision>=3.7.0
motor>=3.6.0
pymongo>=4.9.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.1
